+++ quickstart/kusto_sample_config.json
