+++ quickstart/sample_app.go
@@ -8,10 +8,10 @@ import (
 	"github.com/Azure/azure-kusto-go/kusto/data/errors"
 	"github.com/Azure/azure-kusto-go/kusto/ingest"
 	"github.com/Azure/azure-kusto-go/kusto/kql"
-	"github.com/Azure/azure-kusto-go/kusto/quickstart/utils"
-	"github.com/Azure/azure-kusto-go/kusto/quickstart/utils/authentication"
-	"github.com/Azure/azure-kusto-go/kusto/quickstart/utils/ingestion"
-	"github.com/Azure/azure-kusto-go/kusto/quickstart/utils/queries"
+	"github.com/Azure/azure-kusto-go/quickstart/utils"
+	"github.com/Azure/azure-kusto-go/quickstart/utils/authentication"
+	"github.com/Azure/azure-kusto-go/quickstart/utils/ingestion"
+	"github.com/Azure/azure-kusto-go/quickstart/utils/queries"
 	"io/ioutil"
 	"os"
 )
