+++ quickstart/utils/authentication/authentication.go
@@ -4,7 +4,7 @@ package authentication
 import (
 	"fmt"
 	"github.com/Azure/azure-kusto-go/kusto"
-	"github.com/Azure/azure-kusto-go/kusto/quickstart/utils"
+	"github.com/Azure/azure-kusto-go/quickstart/utils"
 	"os"
 )
